Move vcharts into src directory (alongside candela; outside of external)
