+++ src/candela/components/VCharts.js
@@ -1,5 +1,5 @@
 import VisualizationComponent from '../core/VisualizationComponent';
-import vcharts from '../../external/vcharts/src';
+import vcharts from '../../vcharts/src';
 
 export default class LineChart extends VisualizationComponent {
   constructor (el, data) {
