+++ src/vcharts/.npmignore
