+++ src/vcharts/NOTICE
