+++ src/vcharts/bower.json
