+++ src/vcharts/examples/bar.html
