+++ src/vcharts/examples/box.html
