+++ src/vcharts/examples/bullet.html
