+++ src/vcharts/examples/histogram.html
