+++ src/vcharts/examples/xy.html
