+++ src/vcharts/examples/xy_scaling.html
