+++ src/vcharts/examples/xymatrix.html
