+++ src/vcharts/templates/axis.json
