+++ src/vcharts/templates/bar.json
