+++ src/vcharts/templates/box.json
