+++ src/vcharts/templates/bullet.json
