+++ src/vcharts/templates/gantt.json
