+++ src/vcharts/templates/histogram.json
