+++ src/vcharts/templates/vega.json
