+++ src/vcharts/templates/xy.json
